+++ packages/cli/src/commands/setup-database.ts
@@ -1,29 +1,37 @@
 import { exec } from "node:child_process";
 import { performance } from "node:perf_hooks";
 import { promisify } from "node:util";
-import { convertMeasureToPrettyString, startRunningMessage } from "../utils.js";
+import {
+	convertMeasureToPrettyString,
+	getSmashConfig,
+	startRunningMessage,
+} from "../utils.js";
 import "dotenv/config";
 
 export const command = "setup-database";
 export const describe =
 	"Create a new database and initialise the site with no content.";
-export default async function setupDatabase() {
+export async function handler() {
 	const execute = promisify(exec);
-	const themeName = process.env.npm_package_config_theme_name;
+	const smashConfig = await getSmashConfig();
+	// These must remain env vars because they differ for each dev.
 	const addCustomUser =
 		process.env.WORDPRESS_USER &&
 		process.env.WORDPRESS_USER_EMAIL &&
 		process.env.WORDPRESS_PASSWORD;
 
-	if (!themeName) {
-		throw new Error("Theme name is missing from package.json config object.");
+	if (!smashConfig) {
+		throw new Error(
+			"Unable to determine project setup information. Please add a smash.config.ts file with the required info.",
+		);
 	} else {
-		const interval = startRunningMessage("Initialising database");
+		const { projectName, themeFolderName } = smashConfig;
+		const stopRunningMessage = startRunningMessage("Initialising database");
 		performance.mark("Start");
 		await execute("wp db create")
 			.then(() => {
 				return execute(
-					`wp core install --url=http://${process.env.CI ? "127.0.0.1" : `${themeName}.test`}/ --title=Temp --admin_user=Bot --admin_email=fake@fake.com --admin_password=password`,
+					`wp core install --url=http://${process.env.CI ? "127.0.0.1" : `${projectName}.test`}/ --title=Temp --admin_user=Bot --admin_email=fake@fake.com --admin_password=password`,
 				);
 			})
 			.then(() => {
@@ -62,28 +70,24 @@ export default async function setupDatabase() {
 						),
 					)})`,
 				);
-				return execute(`wp theme activate ${themeName}`);
+				return execute(`wp theme activate ${themeFolderName}`);
 			})
-			.then(() => {
+			.then(async () => {
 				performance.mark("theme");
 				console.log(
 					`Theme activated. (${convertMeasureToPrettyString(
 						performance.measure("theme", "plugins"),
 					)})`,
 				);
-				if (interval !== null) {
-					clearInterval(interval);
-				}
+				await stopRunningMessage();
 				console.log(
 					`Database set up${addCustomUser ? ` and ${process.env.WORDPRESS_USER} user added` : !process.env.CI ? ". To set up a user, run the `wp user create` command." : ""}. (${convertMeasureToPrettyString(
 						performance.measure("everything", "Start"),
 					)})`,
 				);
 			})
-			.catch((error: { stderr: string }) => {
-				if (interval !== null) {
-					clearInterval(interval);
-				}
+			.catch(async (error: { stderr: string }) => {
+				await stopRunningMessage();
 				if (error.stderr?.startsWith("ERROR 1007")) {
 					console.error(
 						"Database already exists with the name in the wp-config. Please delete that database first with `wp db drop --yes`",