+++ packages/test-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atomicsmash/test-utils
 
+## 5.0.4-beta.0
+
+### Patch Changes
+
+- [#445](https://github.com/AtomicSmash/packages/pull/445) [`90629ab`](https://github.com/AtomicSmash/packages/commit/90629aba40fe2e06200a6294f66b870cd60864e5) Thanks [@mikeybinns](https://github.com/mikeybinns)! - Update deps
+
 ## 5.0.3
 
 ### Patch Changes